finish: - add good item - remove good item - edit good item - vote good item

+++ app/Http/Controllers/Retrospective/ItemManager.php
@@ -25,32 +25,41 @@ class ItemManager extends Controller
     public function fetch(Request $request)
     {
         try {
-            $results = [
-                'goods' => [
-                    [
-                        'id' => 7,
-                        'content' => "Some quick example text to build on the card title and make up the bulk of the card's content.",
-                        'vote' => 1,
-                        'isVoted' => false,
-                    ]
-                ],
-                'bads' => [
-                    [
-                        'id' => 1,
-                        'content' => "Some quick example text to build on the card title and make up the bulk of the card's content.",
-                        'vote' => 1,
-                        'isVoted' => false,
-                    ]
-                ],
-                'ideas' => [
-                    [
-                        'id' => 1,
-                        'content' => "Some quick example text to build on the card title and make up the bulk of the card's content.",
-                        'vote' => 1,
-                        'isVoted' => false,
-                    ]
-                ]
-            ];
+            $meetingId = $request->get('meetingId');
+            $attendeeId = $request->get('attendeeId');
+
+            $items = $this->itemRepo->getItems($meetingId);
+            if (empty($items)) {
+                return response()->json([], 200);
+            }
+
+            $votes = $this->itemRepo->getVotes($attendeeId);
+            $userVotes = [];
+            foreach ($votes as $vote) {
+                $userVotes[$vote->item_id] = $vote->attendee_id;
+            }
+
+            $results = [];
+            foreach ($items as $key => $item) {
+                $isVoted = false;
+                if (!empty($userVotes[$item->id]) && $userVotes[$item->id] === $attendeeId) {
+                    $isVoted = true;
+                }
+
+                $ticket = [
+                    'id' => $item->id,
+                    'content' => $item->content,
+                    'isVoted' => $isVoted,
+                    'vote' => $item->vote
+                ];
+                if ($item->type === 0) {
+                    $results['goods'][] = $ticket;
+                } else if ($item->type === 1) {
+                    $results['bads'][] = $ticket;
+                } else if ($item->type === 2) {
+                    $results['ideas'][] = $ticket;
+                }
+            }
 
             return response()->json($results, 200);
 
@@ -62,21 +71,22 @@ class ItemManager extends Controller
     public function add(Request $request)
     {
         try {
-            $attendeeId = $request->get('attendeeId');
-            $meetingId = $request->get('meetingId');
 
-//            if (!$this->meetingRepo->doesMeetingExist($meetingId)){
-//                return response()->json(['message' => 'meeting does not exist'], 500);
-//            }
-//
-//            if (!$this->attendeeRepo->doesAttendeeExist($attendeeId)){
-//                return response()->json(['message' => 'attendee does not exist'], 500);
-//            }
+            $meetingId = $request->get('meetingId');
+            if (!$this->meetingRepo->doesMeetingExist($meetingId)) {
+                return response()->json(['message' => 'meeting does not exist'], 500);
+            }
+
+            $attendeeId = $request->get('attendeeId');
+            if (!$this->attendeeRepo->doesAttendeeExist($attendeeId)) {
+                return response()->json(['message' => 'attendee does not exist'], 500);
+            }
 
             $item = [
                 'meeting_id' => $meetingId,
                 'attendee_id' => $attendeeId,
                 'type' => $request->get('type'),
+                'vote' => 0,
                 'content' => $request->get('content')
             ];
 
@@ -90,20 +100,46 @@ class ItemManager extends Controller
 
     public function vote(Request $request)
     {
+
         try {
-            $results = [];
-            return response()->json($results, 200);
+            $itemId = $request->get('itemId');
+            $attendeeId = $request->get('attendeeId');
+            $value = $request->get('voteValue');
+
+            if ($value === "true") {
+                $value = true;
+            } else {
+                $value = false;
+            }
+
+            $result = $this->itemRepo->vote($itemId, $attendeeId, $value);
+            if ($result === true) {
+                $result = $this->itemRepo->updateTotalVote($itemId, $value);
+            }
+
+            return response()->json($result, 200);
 
         } catch (Exception $e) {
-            return response()->json(['message' => $e->getMessage()], 500);
+            dd($e->getMessage());
+            return response()->json(['message' => 'Something went wrong!'], 500);
         }
     }
 
     public function edit(Request $request)
     {
         try {
-            $results = [];
-            return response()->json($results, 200);
+            $itemId = $request->get('itemId');
+            $attendeeId = $request->get('attendeeId');
+            if (!$this->attendeeRepo->doesAttendeeExist($attendeeId)) {
+                return response()->json(['message' => 'attendee does not exist'], 500);
+            }
+
+            $content = $request->get('content');
+            if (empty($content)) {
+                return response()->json(['message' => 'item content is empty'], 500);
+            }
+            $result = $this->itemRepo->updateItemContent($itemId, $content);
+            return response()->json($result, 200);
 
         } catch (Exception $e) {
             return response()->json(['message' => $e->getMessage()], 500);
@@ -115,12 +151,14 @@ class ItemManager extends Controller
         try {
             $itemId = $request->get('itemId');
             $attendeeId = $request->get('attendeeId');
-
-//            if (!$this->attendeeRepo->doesAttendeeExist($attendeeId)){
-//                return response()->json(['message' => 'attendee does not exist'], 500);
-//            }
+            if (!$this->attendeeRepo->doesAttendeeExist($attendeeId)) {
+                return response()->json(['message' => 'attendee does not exist'], 500);
+            }
 
             $result = $this->itemRepo->remove($itemId);
+            if ($result === 0) {
+                return response()->json(['message' => 'Item does not exist'], 400);
+            }
 
             return response()->json($result, 200);
